Do sbuild inside docker, too
Previously, we used sbuild with a shell executor to build Debian
packages, and built everything else inside a docker container. While
that works, it special-cases Debian and Ubuntu for no particularly good
reason, and makes deploying this on the servers we'll end up using in
the end much more complicated.

Instead, use the tianon/sbuild docker image (which is already set up to
use sbuild for us) to build things. We can then "just" use docker
(almost) everywhere (not on macOS, not on Windows, not with robots).

Will require that tarballs are downloaded from a single server, but we
can perhaps change that later.